Add database query backend and update SQL database
A SQL query can now be directed to the file 'controllers/query_db.php'
for processing. The interface is written with AJAX in mind (request
sended in JSON, result returned as JSON).

The specific form for JSON request is describe within the file.

Also update the SQL file and 'README.md' to includes new changes used by
the vehicle page.